LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/parsing/shape-outside-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: getComputedStyle().shapeOutside</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes/#shape-outside-property">
<meta name="assert" content="shape-outside computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<style>
  #target {
    font-size: 40px;
  }
</style>
<div id="target"></div>
<script>
// TODO: Add inset() tests

test_computed_value("shape-outside", "circle(at calc(75% + 0px) calc(75% + 0px))", "circle(at 75% 75%)");
test_computed_value("shape-outside", "circle(calc(10px + 0.5em) at -50% 50%) border-box", "circle(30px at -50% 50%) border-box");
test_computed_value("shape-outside", "circle(calc(10px - 0.5em) at 50% -50%) border-box", "circle(0px at 50% -50%) border-box");
test_computed_value("shape-outside", "circle(at 10% 20%)");
test_computed_value("shape-outside", "circle(at 10%)", "circle(at 10% 50%)");
test_computed_value("shape-outside", "circle(at 20% 30px)");
test_computed_value("shape-outside", "circle(at 30px center)", "circle(at 30px 50%)");
test_computed_value("shape-outside", "circle(at 40px top)", "circle(at 40px 0%)");
test_computed_value("shape-outside", "circle(at bottom 10% right 20%)", "circle(at 80% 90%)");
test_computed_value("shape-outside", "circle(at bottom calc(10%) right calc(20%))", "circle(at 80% 90%)");
test_computed_value("shape-outside", "circle(at bottom right)", "circle(at 100% 100%)");
test_computed_value("shape-outside", "circle(at calc(10%) calc(20%))", "circle(at 10% 20%)");
test_computed_value("shape-outside", "circle(at center 10em)", "circle(at 50% 400px)");
test_computed_value("shape-outside", "circle(at center 50px)", "circle(at 50% 50px)");
test_computed_value("shape-outside", "circle(at center bottom)", "circle(at 50% 100%)");
test_computed_value("shape-outside", "circle(at center center)", "circle(at 50% 50%)");
test_computed_value("shape-outside", "circle(at center left)", "circle(at 0% 50%)");
test_computed_value("shape-outside", "circle(at center)", "circle(at 50% 50%)");
test_computed_value("shape-outside", "circle(at left bottom)", "circle(at 0% 100%)");
test_computed_value("shape-outside", "circle(at left center)", "circle(at 0% 50%)");
test_computed_value("shape-outside", "circle(at left)", "circle(at 0% 50%)");
test_computed_value("shape-outside", "circle(at right 30% top 60px)", "circle(at 70% 60px)");
test_computed_value("shape-outside", "circle(at right 40%)", "circle(at 100% 40%)");
test_computed_value("shape-outside", "circle(at right bottom)", "circle(at 100% 100%)");
test_computed_value("shape-outside", "circle(at top center)", "circle(at 50% 0%)");
test_computed_value("shape-outside", "circle(at top)", "circle(at 50% 0%)");

test_computed_value("shape-outside", "ellipse(60% closest-side at 50% 50%)");
test_computed_value("shape-outside", "ellipse(calc(10px + 0.5em) calc(10px - 0.5em) at -50% 50%) padding-box", "ellipse(30px 0px at -50% 50%) padding-box");
test_computed_value("shape-outside", "ellipse(calc(10px - 0.5em) calc(10px + 0.5em) at 50% -50%) border-box", "ellipse(0px 30px at 50% -50%) border-box");
test_computed_value("shape-outside", "ellipse(at 10% 20%)");
test_computed_value("shape-outside", "ellipse(at 10%)", "ellipse(at 10% 50%)");
test_computed_value("shape-outside", "ellipse(at 20% 30px)");
test_computed_value("shape-outside", "ellipse(at 30px center)", "ellipse(at 30px 50%)");
test_computed_value("shape-outside", "ellipse(at 40px top)", "ellipse(at 40px 0%)");
test_computed_value("shape-outside", "ellipse(at bottom 10% right 20%)", "ellipse(at 80% 90%)");
test_computed_value("shape-outside", "ellipse(at bottom calc(10%) right calc(20%))", "ellipse(at 80% 90%)");
test_computed_value("shape-outside", "ellipse(at bottom right)", "ellipse(at 100% 100%)");
test_computed_value("shape-outside", "ellipse(at calc(10%) calc(20%))", "ellipse(at 10% 20%)");
test_computed_value("shape-outside", "ellipse(at center 10em)", "ellipse(at 50% 400px)");
test_computed_value("shape-outside", "ellipse(at center 50px)", "ellipse(at 50% 50px)");
test_computed_value("shape-outside", "ellipse(at center bottom)", "ellipse(at 50% 100%)");
test_computed_value("shape-outside", "ellipse(at center center)", "ellipse(at 50% 50%)");
test_computed_value("shape-outside", "ellipse(at center left)", "ellipse(at 0% 50%)");
test_computed_value("shape-outside", "ellipse(at center)", "ellipse(at 50% 50%)");
test_computed_value("shape-outside", "ellipse(at left bottom)", "ellipse(at 0% 100%)");
test_computed_value("shape-outside", "ellipse(at left center)", "ellipse(at 0% 50%)");
test_computed_value("shape-outside", "ellipse(at left)", "ellipse(at 0% 50%)");
test_computed_value("shape-outside", "ellipse(at right 30% top 60px)", "ellipse(at 70% 60px)");
test_computed_value("shape-outside", "ellipse(at right 40%)", "ellipse(at 100% 40%)");
test_computed_value("shape-outside", "ellipse(at right bottom)", "ellipse(at 100% 100%)");
test_computed_value("shape-outside", "ellipse(at top center)", "ellipse(at 50% 0%)");
test_computed_value("shape-outside", "ellipse(at top)", "ellipse(at 50% 0%)");

test_computed_value("polygon(evenodd, -10px, -20px, -30px, -40px, -50px, -60px) margin-box");
test_computed_value("polygon(10%, 20%, 30%, 40%, 50%, 60%) content-box");
test_computed_value("polygon(calc(10px - 0.5em), 20%, 30%, 40%, 50%, calc(10px - 0.5em))", "polygon(-10px, 20%, 30%, 40%, 50%, -10px)");
</script>
</body>
</html>
